requirements.txt
pandas==2.0.3
numpy==1.24.3
sqlalchemy==2.0.19
prophet==1.1.4
streamlit==1.28.0
plotly==5.17.0
seaborn==0.12.2
matplotlib==3.7.2
python-dotenv==1.0.0
psycopg2-binary==2.9.7
sqlite3
scikit-learn==1.3.0